Client game Angelscript code refactoring
